requirements.txt
pandas
openpyxl
mysql-connector-python
gradio
langchain
sqlalchemy
openai
fpdf
